% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char2factor.R
\name{char2factor}
\alias{char2factor}
\title{Convert a Character Vector Into a Factor}
\usage{
char2factor(x, levels = unique(x))
}
\arguments{
\item{x}{A character vector.}

\item{levels}{The levels to use, see \code{\link{factor}}.}
}
\value{
A factor.
}
\description{
I often need a factor with levels the unique values of a character vector
(for example: to prevent ggplot2 from sorting the character vector).
}
\examples{
x <- c("beech", "oak", "spruce", "fir")
char2factor(x)
}
\seealso{
Other vector functions: 
\code{\link{escape_non_ascii}()},
\code{\link{file_string}()},
\code{\link{powers_of_ten}}
}
\concept{vector functions}
