% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_vectors.R
\name{compare_vectors}
\alias{compare_vectors}
\title{Compare Two Vectors}
\usage{
compare_vectors(x, y)
}
\arguments{
\item{x, y}{Two vectors of the same mode.}
}
\value{
A matrix containing the side-by-side comparison.
}
\description{
Side-by-side comparison of two vectors.
}
\examples{
data(mtcars)
cars <- rownames(mtcars)
carz <- cars[-grep("Merc", cars)]
cars <- cars[nchar(cars) < 15]
compare_vectors(cars, carz)
}
