% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_umlauts_to_tex.R
\name{convert_umlauts_to_tex}
\alias{convert_umlauts_to_tex}
\title{Tex Codes for German Umlauts}
\usage{
convert_umlauts_to_tex(x)
}
\arguments{
\item{x}{A string.}
}
\value{
A string with the umlauts converted to plain TeX.
}
\description{
Tex Codes for German Umlauts
}
