% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escape_non_ascii.R
\name{escape_non_ascii}
\alias{escape_non_ascii}
\title{Escape non-ASCII Characters}
\usage{
escape_non_ascii(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector.
}
\description{
I often get code with german umlauts that need to be escaped.
}
\examples{
x <- c("foo", "djörman", "bar", "djörman bar")
escape_non_ascii(x)
# change file
f <- tempfile()
writeLines(x, f)
writeLines(escape_non_ascii(readLines(f)), f)
}
\seealso{
Other vector functions: 
\code{\link{char2factor}()},
\code{\link{file_string}()},
\code{\link{powers_of_ten}}
}
\concept{vector functions}
