% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{get_rscript_script_path}
\alias{get_rscript_script_path}
\title{Get the Path of the R Code File in Case of an Rscript Run}
\usage{
get_rscript_script_path()
}
\value{
A vector of \code{\link{mode}} character giving the name of the R
code file. Will be character(0) if not in an Rscript run.
}
\description{
Get the Path of the R Code File in Case of an Rscript Run
}
\examples{
get_rscript_script_path()
}
