% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unique_string.R
\name{get_unique_string}
\alias{get_unique_string}
\title{Create a Fairly Unique String}
\usage{
get_unique_string()
}
\value{
A fairly unique string.
}
\description{
I sometimes need a fairly unique string, mostly for file names, that should
start with the current date.
}
\examples{
replicate(20, get_unique_string())
}
