% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golden_ratio.R
\name{golden_ratio}
\alias{golden_ratio}
\title{Calculate the Golden Ratio}
\usage{
golden_ratio(x)
}
\arguments{
\item{x}{The sum of the two quantities to be in the golden ratio.}
}
\value{
A numeric vector of length 2, containing the two quantities \emph{a}
and \emph{b}, \emph{a}  being
the larger.
}
\description{
Calculate the Golden Ratio
}
\examples{
golden_ratio(10)
}
