% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{is_batch}
\alias{is_batch}
\title{Is R Run in Batch Mode (via R CMD BATCH or Rscript)?}
\usage{
is_batch()
}
\value{
A boolean.
}
\description{
Is R Run in Batch Mode (via R CMD BATCH or Rscript)?
}
\examples{
is_batch()
}
