% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boolean.R
\name{is_false}
\alias{is_false}
\title{Provide \code{isFALSE} for R < 3.5.0}
\usage{
is_false(x)
}
\arguments{
\item{x}{The object to be tested.}
}
\value{
\code{\link[base:TRUE]{TRUE}} if the object is set to
code{\link[base:FALSE]{FALSE}},
\code{\link[base:FALSE]{FALSE}}
otherwise.
}
\description{
I still use R 3.3.3 for testing, \code{isFALSE()} was introduced in R 3.5.0.
}
\examples{
is_false("not false")
is_false(FALSE)
}
\seealso{
Other logical helpers: 
\code{\link{is_not_false}()},
\code{\link{is_null_or_true}()}
}
\concept{logical helpers}
