% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{is_force}
\alias{is_force}
\title{Opt-out Via Option}
\usage{
is_force(x = .packages()[1])
}
\arguments{
\item{x}{The option under which an element \code{"force"} is to be searched
for.}
}
\value{
\code{\link[base:TRUE]{TRUE}} if option \code{x[["force"]]} is either
\code{\link[base:TRUE]{TRUE}} or \code{\link[base:NULL]{NULL}} (i.e. not set
at all).
}
\description{
Opt-out Via Option
}
\examples{
is_force()
set_options(list(force = FALSE))
get_options(flatten_list = FALSE)
is_force()
}
\seealso{
Other option functions: 
\code{\link{get_options}()},
\code{\link{set_options}()}
}
\concept{option functions}
