% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installed.R
\name{is_installed}
\alias{is_installed}
\title{Is an External Program Installed}
\usage{
is_installed(program)
}
\arguments{
\item{program}{Name of the program.}
}
\value{
\code{\link[base:TRUE]{TRUE}} on success, 
\code{\link[base:FALSE]{FALSE}} otherwise.
}
\description{
Is an External Program Installed
}
\examples{
is_installed("R")
is_installed("probably_not_installed")
}
