% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boolean.R
\name{is_null_or_true}
\alias{is_null_or_true}
\title{Is an Object \emph{TRUE} or  \emph{NULL}}
\usage{
is_null_or_true(x)
}
\arguments{
\item{x}{The object to be tested.}
}
\value{
\code{\link[base:TRUE]{TRUE}} if the object is set to
code{\link[base:TRUE]{TRUE}} or  code{\link[base:NULL]{NULL}},
\code{\link[base:FALSE]{FALSE}}
otherwise.
}
\description{
Is an Object \emph{TRUE} or  \emph{NULL}
}
\examples{
is_false("not false")
is_false(FALSE)
}
\seealso{
\code{\link{is_force}}

Other logical helpers: 
\code{\link{is_false}()},
\code{\link{is_not_false}()}
}
\concept{logical helpers}
