% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_of_length_zero.R
\name{is_of_length_zero}
\alias{is_of_length_zero}
\title{Is an Object of Length Zero?}
\usage{
is_of_length_zero(x, class = NULL)
}
\arguments{
\item{x}{The object.}

\item{class}{An optional character vector of length 1 giving the class. See
\emph{examples}.}
}
\value{
\code{\link[base:TRUE]{TRUE}} on success, 
\code{\link[base:FALSE]{FALSE}} otherwise.
}
\description{
Some expressions evaluate to \code{\link{integer}(0)} or the like.
}
\examples{
x <- ""; length(x); is_of_length_zero(x)
x <- grep(" ", "")
print(x)
is_of_length_zero(x)
is_of_length_zero(x, "character")
is_of_length_zero(x, "numeric")
is_of_length_zero(x, "integer")
}
