% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_r_cmd_check.R
\name{is_r_cmd_check}
\alias{is_r_cmd_check}
\title{Is the Current R Process an \command{R CMD check}?}
\usage{
is_r_cmd_check()
}
\value{
\code{\link[base:TRUE]{TRUE}} on success, 
\code{\link[base:FALSE]{FALSE}} otherwise.
}
\description{
Is the Current R Process an \command{R CMD check}?
}
