% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installed.R
\name{is_r_package_installed}
\alias{is_r_package_installed}
\title{Is an R Package Installed}
\usage{
is_r_package_installed(x, version = "0")
}
\arguments{
\item{x}{Name of the package as character string.}

\item{version}{Required minimum version of the package as character string.}
}
\value{
\code{\link[base:TRUE]{TRUE}} on success, 
\code{\link[base:FALSE]{FALSE}} otherwise.
}
\description{
Is an R Package Installed
}
\examples{
is_r_package_installed("base", "300.0.0")
is_r_package_installed("fritools", "1.0.0")
}
