% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_helpers.R
\name{is_running_on_fvafrcu_machines}
\alias{is_running_on_fvafrcu_machines}
\title{Is the Machine Running the Current R Process Owned by FVAFRCU?}
\usage{
is_running_on_fvafrcu_machines()
}
\value{
\code{\link[base:TRUE]{TRUE}} on success, 
\code{\link[base:FALSE]{FALSE}} otherwise.
}
\description{
Is the Machine Running the Current R Process Owned by FVAFRCU?
}
\seealso{
Other test helpers: 
\code{\link{get_boolean_envvar}()},
\code{\link{get_run_r_tests}()},
\code{\link{is_running_on_gitlab_com}()},
\code{\link{run_r_tests_for_known_hosts}()},
\code{\link{set_run_r_tests}()}
}
\concept{test helpers}
