% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_success.R
\name{is_success}
\alias{is_success}
\title{Does the Return Value of a Command Signal Success?}
\usage{
is_success(x)
}
\arguments{
\item{x}{The external commands return value.}
}
\value{
\code{\link[base:TRUE]{TRUE}} on success, 
\code{\link[base:FALSE]{FALSE}} otherwise.
}
\description{
This is just a wrapper to ease the evaluation of retrun values from external
commands:
External commands return 0 on success, which is
\code{\link[base:FALSE]{FALSE}}, when converted to logical.
}
