% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip_off_attributes.R
\name{strip_off_attributes}
\alias{strip_off_attributes}
\title{Strip Attributes off an Object}
\usage{
strip_off_attributes(x)
}
\arguments{
\item{x}{An object.}
}
\value{
The object.
}
\description{
Strip Attributes off an Object
}
\examples{
y <- stats::setNames(1:3, letters[1:3])
attr(y, "myattr") <- "qwer"
comment(y) <- "qwer"
strip_off_attributes(y)
}
\seealso{
\code{\link[base:unname]{base::unname}}
}
