% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_files.R
\name{summary.filesearch}
\alias{summary.filesearch}
\title{Summarize File Searches}
\usage{
\method{summary}{filesearch}(object, ..., type = c("file", "what", "matches"))
}
\arguments{
\item{object}{A object returned by \code{\link{search_files}}.}

\item{...}{Needed for compatibility.}

\item{type}{Type of summary.}
}
\value{
A summarized object.
}
\description{
Summarize File Searches
}
\seealso{
Other file searching functions: 
\code{\link{find_files}()},
\code{\link{search_files}()}
}
\concept{file searching functions}
