% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_helpers.R
\name{is_running_on_fvafrcu_machines}
\alias{is_running_on_fvafrcu_machines}
\title{Is the Machine Running the Current R Process Owned by FVAFRCU?}
\usage{
is_running_on_fvafrcu_machines()
}
\value{
A logical.
}
\description{
Is the Machine Running the Current R Process Owned by FVAFRCU?
}
\keyword{internal}
