% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_helpers.R
\name{is_running_on_gitlab_com}
\alias{is_running_on_gitlab_com}
\title{Is the Current Machine Owned by \url{https://about.gitlab.com}?}
\usage{
is_running_on_gitlab_com(verbose = TRUE)
}
\arguments{
\item{verbose}{Be verbose?}
}
\value{
\code{\link[base:TRUE]{TRUE}} on success, 
\code{\link[base:FALSE]{FALSE}} otherwise.
}
\description{
Check whether the current machine is located on
\url{https://about.gitlab.com}.
This check is an approximation only.
}
\seealso{
Other test helpers: 
\code{\link{get_run_r_tests}()},
\code{\link{is_running_on_fvafrcu_machines}()},
\code{\link{set_run_r_tests}()}
}
\concept{test helpers}
